git-svn-id: file:///home/behr_mi/git/sumo_synched/branches/emergency_vehicle@19721 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ src/microsim/traffic_lights/MSSimpleTrafficLightLogic.cpp
@@ -38,6 +38,7 @@
 #include <bitset>
 #include <sstream>
 #include <microsim/MSEventControl.h>
+#include <microsim/MSNet.h>
 #include "MSTLLogicControl.h"
 #include "MSTrafficLightLogic.h"
 #include "MSSimpleTrafficLightLogic.h"
@@ -100,9 +101,9 @@ MSSimpleTrafficLightLogic::trySwitch() {
 
 
 // ------------ Static Information Retrieval
-unsigned int
+int
 MSSimpleTrafficLightLogic::getPhaseNumber() const {
-    return (unsigned int) myPhases.size();
+    return (int) myPhases.size();
 }
 
 
@@ -119,14 +120,14 @@ MSSimpleTrafficLightLogic::getPhases() {
 
 
 const MSPhaseDefinition&
-MSSimpleTrafficLightLogic::getPhase(unsigned int givenStep) const {
-    assert(myPhases.size() > givenStep);
+MSSimpleTrafficLightLogic::getPhase(int givenStep) const {
+    assert((int)myPhases.size() > givenStep);
     return *myPhases[givenStep];
 }
 
 
 // ------------ Dynamic Information Retrieval
-unsigned int
+int
 MSSimpleTrafficLightLogic::getCurrentPhaseIndex() const {
     return myStep;
 }
@@ -207,8 +208,8 @@ MSSimpleTrafficLightLogic::changeStepAndDuration(MSTLLogicControl& tlcontrol,
 
 
 void
-MSSimpleTrafficLightLogic::setPhases(const Phases& phases, unsigned int step) {
-    assert(step < phases.size());
+MSSimpleTrafficLightLogic::setPhases(const Phases& phases, int step) {
+    assert(step < (int)phases.size());
     deletePhases();
     myPhases = phases;
     myStep = step;
